add getDownloadLinks method to parse.ts

+++ build/bot.js
@@ -11,11 +11,21 @@ bot.command("start", (ctx) => {
 bot.on("message:text", async (ctx) => {
     const msg = await ctx.reply("Searching...");
     const sites = await parse.searchSites(ctx.message.text);
-    const inlineKeyboard = new grammy.InlineKeyboard();
-    for (let i of sites) {
-        inlineKeyboard.text(i.name, i.url);
+    if (sites.length === 0) {
+        ctx.api.editMessageText(ctx.chat.id, msg.message_id, "Can not find movie");
     }
-    ctx.api.deleteMessage(ctx.chat.id, msg.message_id);
-    ctx.reply("Choose a site for download:", { reply_markup: inlineKeyboard });
+    else {
+        const inlineKeyboard = new grammy.InlineKeyboard();
+        for (let i of sites) {
+            inlineKeyboard.text(i.name, "url:" + i.url);
+        }
+        await ctx.reply("Choose a site for download:", { reply_markup: inlineKeyboard });
+        ctx.api.deleteMessage(ctx.chat.id, msg.message_id);
+    }
+});
+bot.callbackQuery(/url:(.+)/, async (ctx) => {
+    const url = ctx.callbackQuery.data.slice(4);
+    const links = await parse.getDownloadLinks(url);
+    ctx.answerCallbackQuery();
 });
 bot.start();
