+++ source/bot.ts
@@ -18,15 +18,27 @@ bot.on("message:text", async (ctx) => {
 
     const sites = await parse.searchSites(ctx.message.text);
 
-    const inlineKeyboard = new grammy.InlineKeyboard();
+    if (sites.length === 0) {
+        ctx.api.editMessageText(ctx.chat.id, msg.message_id, "Can not find movie");
+    } else {
+        const inlineKeyboard = new grammy.InlineKeyboard();
 
-    for (let i of sites) {
-        inlineKeyboard.text(i.name, i.url);
+        for (let i of sites) {
+            inlineKeyboard.text(i.name, "url:" + i.url);
+        }
+
+        await ctx.reply("Choose a site for download:", { reply_markup: inlineKeyboard });
+
+        ctx.api.deleteMessage(ctx.chat.id, msg.message_id);
     }
-
-    ctx.api.deleteMessage(ctx.chat.id, msg.message_id);
-
-    ctx.reply("Choose a site for download:", { reply_markup: inlineKeyboard });
 });
+
+bot.callbackQuery(/url:(.+)/, async (ctx) => {
+    const url = ctx.callbackQuery.data.slice(4);
+
+    const links = await parse.getDownloadLinks(url);
+
+    ctx.answerCallbackQuery();
+})
 
 bot.start();
